+++ tablesalt/lib/tablesalt.rb
@@ -3,13 +3,13 @@
 require "active_support"
 require "short_circu_it"
 
-require "tablesalt/version"
+require_relative "tablesalt/version"
 
-require "tablesalt/class_pass"
-require "tablesalt/dsl_accessor"
-require "tablesalt/isolation"
-require "tablesalt/stringable_object"
-require "tablesalt/thread_accessor"
-require "tablesalt/uses_hash_for_equality"
+require_relative "tablesalt/class_pass"
+require_relative "tablesalt/dsl_accessor"
+require_relative "tablesalt/isolation"
+require_relative "tablesalt/stringable_object"
+require_relative "tablesalt/thread_accessor"
+require_relative "tablesalt/uses_hash_for_equality"
 
 module Tablesalt; end