+++ tablesalt/lib/tablesalt/rspec/config.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.configure do |config|
+  config.after(:each) do
+    Thread.current[Tablesalt::ThreadAccessor::THREAD_ACCESSOR_STORE_THREAD_KEY] = nil
+  end
+end