+++ tablesalt/lib/tablesalt/spec_helper.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
 require_relative "rspec/custom_matchers"
+require_relative "rspec/config"