+++ tablesalt/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Upcoming <!-- Add unreleased change notes here: -->
 
+## v0.27.0
+
+*Release Date*: 8/23/2021
+
+### ThreadAccessor
+- Isolated ThreadAccessor store, Rack middleware ([#462](https://github.com/Freshly/spicerack/pull/462))
+- Added namespaced thread stores for gem usage ([#463](https://github.com/Freshly/spicerack/pull/463), ([#464](https://github.com/Freshly/spicerack/pull/464)), ([#466](https://github.com/Freshly/spicerack/pull/466)))
+- Add `define_thread_writer` and `define_thread_accessor` custom matchers ([#467](https://github.com/Freshly/spicerack/pull/467))
+
 ## v0.26.0.4
 
 *Release Date*: 5/17/2021